gateway 관리자연결 / message / screen 제어 영역으로 분리
