+++ apps/overlay/src/app/app.gateway.ts
@@ -12,14 +12,8 @@ import { Logger } from '@nestjs/common';
 import {
   SocketInfo,
   SocketIdandDevice,
-  PurchaseMessage,
   PageUrlAndDevice,
-  RoomAndText,
-  RoomAndDate,
-  RoomAndVideoType,
 } from '@project-lc/shared-types';
-import { OverlayService } from '@project-lc/nest-modules';
-import { AppService } from './app.service';
 
 @WebSocketGateway({ cors: true, transports: ['websocket'] })
 export class AppGateway
@@ -29,11 +23,6 @@ export class AppGateway
   server: Server;
 
   socketInfo: SocketInfo = {};
-
-  constructor(
-    private readonly appService: AppService,
-    private readonly overlayService: OverlayService,
-  ) {}
 
   private logger: Logger = new Logger('AppGateway');
 
@@ -105,140 +94,5 @@ export class AppGateway
           fullUrl[0] ? this.socketInfo[fullUrl[0]] : null,
         );
     }
-  }
-
-  @SubscribeMessage('get ranking')
-  async getRanking(@MessageBody() roomName: string) {
-    const nicknameAndPrice = [];
-    const rankings = await this.appService.getRanking();
-    rankings.forEach((eachNickname) => {
-      const price = Object.values(eachNickname._sum).toString();
-      const { nickname } = eachNickname;
-      nicknameAndPrice.push({ nickname, price: Number(price) });
-    });
-    nicknameAndPrice.sort((a, b) => {
-      return b.price - a.price;
-    });
-
-    this.server.to(roomName).emit('get top-left ranking', nicknameAndPrice);
-  }
-
-  @SubscribeMessage('right top purchase message')
-  async getRightTopPurchaseMessage(@MessageBody() purchase: PurchaseMessage) {
-    const { roomName } = purchase;
-    const nicknameAndPrice = [];
-    const bottomAreaTextAndNickname: string[] = [];
-    const rankings = await this.appService.getRanking();
-    const totalSold = await this.appService.getTotalSoldPrice();
-    const messageAndNickname = await this.appService.getMessageAndNickname();
-
-    rankings.forEach((eachNickname) => {
-      const price = Object.values(eachNickname._sum).toString();
-      const { nickname } = eachNickname;
-      nicknameAndPrice.push({ nickname, price: Number(price) });
-    });
-    nicknameAndPrice.sort((a, b) => {
-      return b.price - a.price;
-    });
-
-    messageAndNickname.forEach((d: { nickname: string; text: string }) => {
-      bottomAreaTextAndNickname.push(`${d.text} - [${d.nickname}]`);
-    });
-
-    const audioBuffer = await this.overlayService.googleTextToSpeech(purchase);
-
-    this.server
-      .to(roomName)
-      .emit('get right-top purchase message', [purchase, audioBuffer]);
-    this.server.to(roomName).emit('get top-left ranking', nicknameAndPrice);
-    this.server.to(roomName).emit('get current quantity', totalSold);
-    this.server
-      .to(roomName)
-      .emit('get bottom purchase message', bottomAreaTextAndNickname);
-  }
-
-  @SubscribeMessage('get all data')
-  async getAllPurchaseMessage(@MessageBody() roomName: string) {
-    const nicknameAndPrice = [];
-    const bottomAreaTextAndNickname: string[] = [];
-    const rankings = await this.appService.getRanking();
-    const totalSold = await this.appService.getTotalSoldPrice();
-    const messageAndNickname = await this.appService.getMessageAndNickname();
-
-    rankings.forEach((eachNickname) => {
-      const price = Object.values(eachNickname._sum).toString();
-      const { nickname } = eachNickname;
-      nicknameAndPrice.push({ nickname, price: Number(price) });
-    });
-    nicknameAndPrice.sort((a, b) => {
-      return b.price - a.price;
-    });
-
-    messageAndNickname.forEach((d: { nickname: string; text: string }) => {
-      bottomAreaTextAndNickname.push(`${d.text} - [${d.nickname}]`);
-    });
-
-    this.server.to(roomName).emit('get top-left ranking', nicknameAndPrice);
-    this.server.to(roomName).emit('get current quantity', totalSold);
-    this.server
-      .to(roomName)
-      .emit('get bottom purchase message', bottomAreaTextAndNickname);
-  }
-
-  @SubscribeMessage('toggle right-top onad logo')
-  handleLogo(@MessageBody() roomName: string) {
-    this.server.to(roomName).emit('toggle right-top onad logo from server');
-  }
-
-  @SubscribeMessage('bottom area message')
-  handleBottomAreaReply(@MessageBody() data: RoomAndText) {
-    const { roomName } = data;
-    const { message } = data;
-    this.server.to(roomName).emit('get bottom area message', message);
-  }
-
-  @SubscribeMessage('toggle bottom area from admin')
-  handleBottomMessageArea(@MessageBody() roomName: string) {
-    this.server.to(roomName).emit('handle bottom area to client');
-  }
-
-  @SubscribeMessage('show live commerce')
-  showLiveCommerce(@MessageBody() roomName: string) {
-    this.server.to(roomName).emit('show screen');
-  }
-
-  @SubscribeMessage('quit live commerce')
-  hideLiveCommerce(@MessageBody() roomName: string) {
-    this.server.to(roomName).emit('hide screen');
-  }
-
-  @SubscribeMessage('get d-day')
-  getDday(@MessageBody() roomAndDate: RoomAndDate) {
-    const { date } = roomAndDate;
-    const { roomName } = roomAndDate;
-    this.server.to(roomName).emit('d-day from server', date);
-  }
-
-  @SubscribeMessage('get non client purchase message from admin')
-  getNonClientMessage(@MessageBody() data: PurchaseMessage) {
-    const { roomName } = data;
-    this.server.to(roomName).emit('get non client purchase message', data);
-  }
-
-  @SubscribeMessage('refresh')
-  handleRefresh(@MessageBody() roomName: string) {
-    this.server.to(roomName).emit('refresh signal');
-  }
-
-  @SubscribeMessage('show video from admin')
-  showVideo(@MessageBody() roomAndType: RoomAndVideoType) {
-    const { roomName } = roomAndType;
-    const { type } = roomAndType;
-    this.server.to(roomName).emit('show video from server', type);
-  }
-
-  @SubscribeMessage('clear full video')
-  clearVideo(@MessageBody() roomName: string) {
-    this.server.to(roomName).emit('clear full video from server');
   }
 }
